backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
yt-dlp==2025.3.31
python-dotenv==1.1.0
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
